chore(frontend): clean after merge
