--- frontend/colab_team_frontend/src/actions/authActions.ts
@@ -1,3 +1,0 @@
-export const authActions = () => {
-  type: "TOGGLE_PASSWORD_VISIBILITY";
-};
