--- frontend/colab_team_frontend/src/components/auth/Container.tsx
@@ -1,5 +1,0 @@
-import Wave from "@/assets/svg/Wave";
-
-export default function SimpleContainer() {
-  return <Wave />;
-}
