+++ frontend/colab_team_frontend/src/components/home/About.tsx
@@ -1,4 +1,5 @@
 import { Box, Container, Typography } from "@mui/material";
+
 import ColabAbout from "@/assets/images/Colab About Light.png";
 import Image from "@/components/Image";
 import { AboutBoxStyle } from "@/styles/appStyles/HomeStyle";
