+++ frontend/colab_team_frontend/src/components/navbar/links.ts
@@ -1,0 +1,9 @@
+import { ABOUT, DISCUSS, PROJECTS, WORKSPACE } from "@/api/app/appEndpoints";
+
+export const navLinks = [
+  { label: "Home", href: "/" },
+  { label: "Projects", href: PROJECTS },
+  { label: "Workspace", href: WORKSPACE },
+  { label: "Discuss", href: DISCUSS },
+  { label: "About", href: ABOUT },
+];
