+++ frontend/colab_team_frontend/src/components/project/RolesAndResponsiblities.tsx
@@ -1,6 +1,4 @@
-import React from "react";
-
-import { Box, Typography, Container } from "@mui/material";
+import { Box, Container, Typography } from "@mui/material";
 
 export default function RolesAndResponsiblities() {
   return (
