+++ frontend/colab_team_frontend/src/layouts/AppContent.tsx
@@ -1,9 +1,8 @@
 import { Box } from "@mui/material";
+import { PropsWithChildren } from "react";
 
 import { AppContentLayoutStyle } from "@/styles/appStyles/AppContentLayout";
 
-export default function AppContentLayout({
-  children,
-}: React.PropsWithChildren) {
+export default function AppContentLayout({ children }: PropsWithChildren) {
   return <Box sx={AppContentLayoutStyle}>{children}</Box>;
 }
