+++ frontend/colab_team_frontend/src/main.tsx
@@ -1,15 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+
 import App from "./App.js";
 
-import { ThemeProvider } from "@mui/material";
-
-import { GlobalTheme } from "./styles/GlobalStyles.js";
-
-ReactDOM.createRoot(document.getElementById("root")!).render(
-  <React.StrictMode>
-    <ThemeProvider theme={GlobalTheme}>
-      <App />
-    </ThemeProvider>
-  </React.StrictMode>
+createRoot(document.querySelector("#root")!).render(
+  <StrictMode>
+    <App />
+  </StrictMode>
 );
