--- frontend/colab_team_frontend/src/states/authInitialState.ts
@@ -1,6 +1,0 @@
-import { ReducerAppStateTypes } from "../typings/authTypes";
-
-export const AuthInitialState: ReducerAppStateTypes = {
-  continueRegistrationForm: false,
-  togglePasswordView: false,
-};
