+++ frontend/colab_team_frontend/src/styles/appStyles/ProjectListStyles.ts
@@ -20,4 +20,4 @@ const ProjectCardContentStyle = {
   },
 };
 
-export { ProjectListCardStyle, ProjectCardContentStyle };
+export { ProjectCardContentStyle,ProjectListCardStyle };
